+++ helpers/index.js
@@ -3,3 +3,4 @@ export { listRewards } from './dm_reward.js';
 export { mapToDisplayStr, formatJSDate } from './formatting.js';
 export { displayAuctionShort, displayAuctionDetails, displayAuctionList, placeBid } from './auction.js';
 export { authorIdFilterFactory } from './generic_discord.js';
+export { displayCharDetails, displayCharList } from './character.js';